Fix method hooking reliability
By deoptimizing everything when the first method is hooked.

This is essential on newer versions of Android where optimizations are
more aggressive. It is however a performance regression, but we can
always optimize this later by hooking the optimized native code.

Kudos to @giantpune for helping track this one down.