Rewrite Java.deoptimizeEverything() using art::Dbg
Accomplishing this using the underlying Instrumentation API directly
means we'd have to deal with a lot of internals to be able to do so
safely. It also makes it tricky in case a debugger is already present.

Using the higher level Dbg API sidesteps these issues, at the expense
of being a bit more intrusive considering the listening socket being
opened. But we can always instrument that away later.