Fix invalid instruction crash on arm64
The logic that recompiles ExceptionClear() did not stop parsing after
the noreturn call to __stack_chk_fail(), which typically falls through
to the next function. That meant we recompiled a lot more code than we
needed to. That was however the best-case scenario, because turns out
there was another scenario where alignment padding was inserted before
the next function, and Instruction.parse() would choke on that.

The ideal solution would have been to recognize the call to
__stack_chk_fail(), but given how complex that is, we instead look out
for a sudden fall-through to the next function, or an unexpected zero-
padding word.

Kudos to @giantpune for helping track this one down.