Speed up the GetOatQuickMethodHeader() hook
Turns out this function is quite hot when running the VM in interpreter-
only mode. E.g. it's called close to 8 million times when launching the
Dropbox app from start to login screen.

We speed things up by emitting a little bit of machine code that looks
at the ArtMethod and avoids the context-switch into JS for methods that
are definitely not hooked.