Revive multi-instance support
There could be more than one script using frida-java, so we cannot
assume that:

A) We know of all the hooked methods.
B) GetOatQuickMethodHeader() hasn't already been replaced.

The challenge with A) is that ArtMethod doesn't have any unused bits we
can use to tag the patched instances.

We mitigate A) by detecting the entrypoint of Frida's libffi closures,
which is going to be the same assuming agent and gadget aren't loaded at
the same time and both using frida-java.

For now we gracefully handle B) by assuming that the first script stays
around for as long as the other scripts depending on this. It's not ideal,
but it will have to do for now.