Java.use()/Java.classFactory.use(): add option to bypass caching
This patch introduces the ability to bypass the "usedClass" caching
implementation of the class-factory.js class loader. Prior to this patch, if
two separate java.lang.ClassLoader instances were to load the same class name
string, even if one were to override Java.classFactory.loader prior to each
Java.classFactory.use() call, all calls to use() after the first would result
in the cached (generated) kclass resolver function being used to load the
class, skipping the current loader entirely. For consistency, this patch also
disables calls to setUsedClass() to prevent unexpected side effects. While
it's clearly not performant to be taking the long route each time, there
appears to be no other way to deal with this issue currently as the
"usedClass" mapping is based only on the class name and not also on the
loader.