Make method interception safer with VMThread
Install new methods using VMThread::execute which blocks all Java
threads and makes it safer to do interception of hot methods.

However, compiled methods that are on the stack will still not be
flushed when calling the deoptimization logic. This is a problem
when reverting back to the old method and flushing the new method
which has been allocated using Frida’s Memory module. In this instance
we force the flushing of old compiled methods using the
NMethodSweeper::force_sweep API. We have to do it twice because of
some logic in nmethod::can_convert_to_zombie.