Add support for opaque JNI IDs on Android R
The JNI IDs for methods and fields should now be treated as opaque. ART
introduced this change in 21d5994583c679cd5d8573b5d35dbd659bdca2c7.

The indirection mode affects method hooks when enabled for the ART
runtime since the method ID returned from JNI function calls will be set
to a managed ID instead of the “real” one that points to the ArtMethod.

This mode will typically be enabled if the application has the
“debuggable” attribute set to true.