Fix ART method hooking reliability after GC (#204)
Since the replacement method that gets allocated by us is not visible to
the ART runtime, it therefore cannot be candidate for garbage collector
visits when sweeps occur. In that case, the declaring_class_ field of
ArtMethod (a GC root) could become stale after it gets moved to another
memory space - resulting in a crash when it is used after the fact.

Fixes #205.