Fix Java.array() for unloaded array types
If the type of the array was not used in the application it means that
it wasn't loaded in memory and Java.array() failed because of this. This
fix tries first to use the array type and if it fails, it will first
create an empty array of that type using
java.lang.reflect.Array.newInstance() before calling factory.use().