jvm: Add preliminary support for OpenJDK 17 on Windows (#328)
This change will make the jvm.js code be able to bridge Java OpenJDK 17
on Windows. This was tested on Windows 10, using jdk-17.0.11+9 from
https://learn.microsoft.com/en-gb/java/openjdk/download and the
separately downloaded and extracted debug symbols.

The code is not pretty, but I've tried to keep it as close to the
original as possible.

Note that I discovered a bug while performing the tests;
https://github.com/frida/frida-gum/issues/811. This means that the
target process needs to be restarted after detaching Frida, for the Java
bridge (resolving of Symbols) to work.

This code currently does not work with OpenJDK21!

This code requires the 'manual' placement of the debug symbols,
otherwise jvm.js will be unable to find the functions it requires.